first commit / validate env
