+++ backend/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "dotenv": "^16.3.1",
+    "envalid": "^8.0.0",
     "express": "^4.18.2",
     "mongoose": "^8.0.4",
     "nodemon": "^3.0.2"
